fix(gateway): apply event loop drain to all CLI gateway calls
Move waitForEventLoopReady into a shared module (event-loop-ready.ts)
and call it in executeGatewayRequestWithScopes in addition to
probeGateway.  This fixes commands like `cron list`, `devices list`,
and any other CLI path that goes through callGateway — they hit the
same deferred ESM module evaluation stall that was causing probe
timeouts.
